requirements.txt
PyGitHub
click
